+++ metacar_hybrid/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.1.0a4"
+__version__ = "0.1.0a5"
 
 from .sceneapi import SceneAPI
 from .geometry import Vector2, Vector3